cleaned up old localization scripts
