+++ src/kuka_kickass_kalman/scripts/kf_node.py
@@ -4,7 +4,7 @@ import time
 from std_msgs.msg import String
 from geometry_msgs.msg import Twist
 from geometry_msgs.msg import Pose
-from ros_kalman_filter2 import kalman_filter
+from ros_kalman_filter import kalman_filter
 from kuka_kickass_kalman.msg import Obs
